moved requests controller to chairs controller
