+++ app/models/ability.rb
@@ -32,13 +32,16 @@ class Ability
   end
 
   def initialize_representative(user)
-    can :show, Chair
+    can :read,      Chair
+    can :requests,  Chair do |chair|
+      user.is_representative?(chair)
+    end
     #can accept, Anträge
     #can show, Holidays of chair members
   end
 
   def initialize_admin(user)
-    can :show,              Chair
+    can :read,              Chair
     can :accept_request,    Chair
     can :remove_from_chair, Chair
     can :set_admin,         Chair
