+++ app/models/chair.rb
@@ -11,6 +11,8 @@
 #
 
 class Chair < ActiveRecord::Base
+  include Rails.application.routes.url_helpers
+
   has_many :chair_wimis, dependent: :destroy
   has_many :users, through: :chair_wimis
   has_many :projects
@@ -97,5 +99,28 @@ class Chair < ActiveRecord::Base
       end
     end
     return success
+  end
+
+  def get_all_requests
+    allrequests = Array.new
+
+    users.each do |user|
+      user.holidays.each do |holidays|
+        unless holidays.status == 'saved'
+          allrequests << {:name => holidays.user.name, :type => 'Holiday Request', :handed_in => holidays.created_at, :status => holidays.status, :action => holiday_path(holidays)}
+        end
+      end
+      user.expenses.each do |expense|
+        unless expense.status == 'saved'
+          allrequests << {:name => expense.user.name, :type => 'Expense Request', :handed_in => expense.created_at, :status => expense.status, :action => expense_path(expense)}
+        end
+      end
+      user.trips.each do |trips|
+        unless trips.status == 'saved'
+          allrequests << {:name => trips.user.name, :type => 'Trip Request', :handed_in => trips.created_at, :status => trips.status, :action => trip_path(trips)}
+        end
+      end
+    end
+    return allrequests.sort_by { |v| v[:handed_in] }.reverse
   end
 end
