+++ config/routes.rb
@@ -21,15 +21,14 @@ Rails.application.routes.draw do
     end
   end
   resources :chairs
-  
+
   post 'chairs/apply', to: 'chairs#apply'
   post 'chairs/accept', to: 'chairs#accept_request'
   post 'chairs/remove_user', to: 'chairs#remove_from_chair'
   post 'chairs/destroy', to: 'chairs#destroy'
   post 'chairs/set_admin', to: 'chairs#set_admin'
   post 'chairs/withdraw_admin', to: 'chairs#withdraw_admin'
-
-  get 'chairs/:id/requests' => 'requests#requests', :as => 'requests'
+  get 'chairs/:id/requests' => 'chairs#requests'
 
   get 'projects/typeahead/:query' => 'projects#typeahead'
